# public/api-docs/News Aggregator API  openapi .yaml
openapi: 3.0.0
info:
  title: News Aggregator API
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
tags:
  - name: Auth
  - name: User Preference
  - name: Article
paths:
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: User Register
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: Suba
                email: subho@test.lk
                password: Test@123344
                password_confirmation: Test@123344
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: User Login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                email: Subashini@test.lk
                password: Test@123344
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: User Logout
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example: ''
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Forget Password
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                email: subho@test.lk
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/auth/reset-password:
    post:
      tags:
        - Auth
      summary: Reset Password
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                token: SHnfCWaX2WtCBI3Zf4hdgjirbHbMLVOcjX5LG89Da84bf090
                password: Test@123344
                password_confirmation: Test@123344
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/user/preferences:
    post:
      tags:
        - User Preference
      summary: Add User Preference
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                preferred_sources:
                  - newsapi
                preferred_categories:
                  - sports
                preferred_authors: []
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    get:
      tags:
        - User Preference
      summary: View User Preference
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/articles/user-preferences:
    get:
      tags:
        - Article
      summary: Get Article with User Preference
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/articles/show:
    get:
      tags:
        - Article
      summary: View Article
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          schema:
            type: integer
          example: '1'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/articles/search:
    get:
      tags:
        - Article
      summary: Search Article
      security:
        - bearerAuth: []
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
          example: Test
        - name: fromDate
          in: query
          schema:
            type: string
          example: '2025-01-10'
        - name: toDate
          in: query
          schema:
            type: string
          example: '2025-01-15'
        - name: category
          in: query
          schema:
            type: string
          example: technology
        - name: source
          in: query
          schema:
            type: string
          example: guardian
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/articles/list:
    get:
      tags:
        - Article
      summary: List Article
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}